requirements.txt
PyDrive
pyinstaller
pyyaml
pytest